+++ src/engine/Code/include/relboost/tests/test10_same_units_categorical.hpp
@@ -66,7 +66,7 @@ void test10_same_units_categorical()
 
     for ( size_t i = 0; i < join_keys_population.size(); ++i )
         {
-            join_keys_population[i] = i;
+            join_keys_population[i] = static_cast<std::int32_t>( i );
         }
 
     const auto join_keys_population_col =
