+++ src/engine/Code/include/relboost/tests/test16_nan_values_numerical.hpp
@@ -54,7 +54,7 @@ void test16_nan_values_numerical()
 
     for ( size_t i = 0; i < join_keys_population.size(); ++i )
         {
-            join_keys_population[i] = i;
+            join_keys_population[i] = static_cast<std::int32_t>( i );
         }
 
     const auto join_keys_population_col =
